// src/app/client/models/restaurant.request.ts
export interface RestauntRequest {
    name: string | undefined;
    city: string | undefined;
    address: string | undefined;
    phone: string | undefined;
    type: string | undefined;
    email: string | undefined;
    description: string | undefined;
    openingHours: string | undefined;
    closingHours: string | undefined;
    capacity: number | undefined;
  }
  